update kubernetes version and isolate e2e test
