+++ scripts/run-e2e-test.sh
@@ -14,7 +14,7 @@ export GO111MODULE=on
 export ACK_GINKGO_DEPRECATIONS=1.16.4
 export ACK_GINKGO_RC=true
 
-if ! (go mod vendor && go get -u github.com/onsi/ginkgo/ginkgo); then
+if ! (go get -u github.com/onsi/ginkgo/ginkgo); then
     echo "go mod vendor or go get ginkgo error"
     exit 1
 fi
@@ -29,7 +29,7 @@ else
     read -ra OPTS <<<"-v $GINKGO_OPTS"
 fi
 
-ginkgo -mod=mod "${OPTS[@]}" tests/e2e
+cd tests/e2e && ginkgo "${OPTS[@]}"
 
 # Checking for test status
 TEST_PASS=$?
