+++ tests/e2e/e2e_test.go
@@ -1,10 +1,8 @@
 package e2e
 
 import (
-	"flag"
 	"fmt"
 	"os"
-	"path/filepath"
 	"testing"
 	"time"
 
@@ -18,7 +16,6 @@ import (
 	. "github.com/onsi/ginkgo"
 	. "github.com/onsi/gomega"
 	"k8s.io/kubernetes/test/e2e/framework"
-	k8sConfig "k8s.io/kubernetes/test/e2e/framework/config"
 )
 
 const (
@@ -32,8 +29,6 @@ var (
 	beautify      string
 )
 
-const kubeconfigEnvVar = "KUBECONFIG"
-
 // TestE2E -
 func TestE2E(t *testing.T) {
 	if testing.Short() {
@@ -43,20 +38,6 @@ func TestE2E(t *testing.T) {
 	initializeFramework()
 	RegisterFailHandler(Fail)
 	RunSpecs(t, "CSM Operator End-to-End Tests")
-}
-
-func initializeFramework() {
-	// k8s.io/kubernetes/tests/e2e/framework requires env KUBECONFIG to be set
-	// it does not fall back to defaults
-	if os.Getenv(kubeconfigEnvVar) == "" {
-		kubeconfig := filepath.Join(os.Getenv("HOME"), ".kube", "config")
-		os.Setenv(kubeconfigEnvVar, kubeconfig)
-	}
-	framework.AfterReadingAllFlags(&framework.TestContext)
-
-	k8sConfig.CopyFlags(k8sConfig.Flags, flag.CommandLine)
-	framework.RegisterCommonFlags(flag.CommandLine)
-	flag.Parse()
 }
 
 var _ = BeforeSuite(func() {
